+++ pkgs/gemini_box/lib/src/freezed/request_body/safety_setting/safety_setting.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: constant_identifier_names
-
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:gemini_box/src/freezed/enums/harm_block_threshold.dart';
 import 'package:gemini_box/src/freezed/enums/harm_category.dart';